#33 Do not use __new__ anymore
You cannot pass an expression or a string to the Expression constructor
anymore. Instead you build an Expression (which is really akin to an
algebra definition) and you can build expression with the configured
operations, symbol, base elements, etc. Or invoke `parse` on it to parse
from a string.

Creating an expression no longer invoke `simplify()` by default: it
needs to be called explicitly, though some method still retain a
simplify arg for convenience.

You can configure an Expression by providing alternative classes for
AND,OR,NOT,TRUE,FALSE,Symbol and the tokenizer function. In particular
you can subclass and customize the way expressions are printed or
represented.This supports #36

`parse`, `normalize`, `symbols` (now `build_symbols`) are methods of the
Expression object. `parse` now uses the symbol class of the configured
expression.

Several attributes shielded by a property are now just simple
attributes. TRUE and FALSE are no longer singletons: instead you can
instantiate them as needed.

A new `.pretty()` method is available and return a pretty formatted
representation of an Expression.


Some internal attributes or methods were renamed for clarity:
 -_cls_order is now `sort_order`.
 - References to cmp have been replaced by comparator.
 - `.remove` has been renamed to `.subtract` to avoid ambiguities with
the list.remove calls.